camera: Renamed "video_capture" files to "camera" and moved to own subdir.
