+++ include/SDL3/SDL.h
@@ -35,6 +35,7 @@
 #include <SDL3/SDL_audio.h>
 #include <SDL3/SDL_bits.h>
 #include <SDL3/SDL_blendmode.h>
+#include <SDL3/SDL_camera.h>
 #include <SDL3/SDL_clipboard.h>
 #include <SDL3/SDL_cpuinfo.h>
 #include <SDL3/SDL_endian.h>
@@ -76,7 +77,6 @@
 #include <SDL3/SDL_touch.h>
 #include <SDL3/SDL_version.h>
 #include <SDL3/SDL_video.h>
-#include "SDL3/SDL_video_capture.h"
 #include <SDL3/SDL_oldnames.h>
 
 #endif /* SDL_h_ */
