+++ src/camera/SDL_camera_c.h
@@ -19,7 +19,7 @@
   3. This notice may not be removed or altered from any source distribution.
 */
 #include "../SDL_internal.h"
-#include "../../include/SDL3/SDL_video_capture.h"
+#include "../../include/SDL3/SDL_camera.h"
 
 #ifndef SDL_video_capture_c_h_
 #define SDL_video_capture_c_h_
