+++ src/camera/SDL_syscamera.h
