+++ src/camera/android/SDL_camera_android.c
