+++ test/testcamera.c
